Restructure tests on Web and Mobile.
