+++ src/main/java/pages/CareersPage.java
@@ -1,4 +1,4 @@
-package pages.web;
+package pages;
 import com.beust.ah.A;
 import com.microsoft.playwright.ElementHandle;
 import com.microsoft.playwright.Page;
